+++ src/Oro/Bundle/GoogleTagManagerBundle/Resources/public/js/app/components/analytics4/home-page-slider-gtm-analytics4-component.js
@@ -1,17 +1,17 @@
 define(function(require) {
-    const BaseComponent = require('orogoogletagmanager/js/app/components/base-embedded-list-gtm-component');
+    const BaseEmbeddedListGtmComponent = require('orogoogletagmanager/js/app/components/base-embedded-list-gtm-component');
     const mediator = require('oroui/js/mediator');
     const _ = require('underscore');
 
     /**
      * Listens to oro:embedded-list:* events and invokes promo "select_promotion, "view_promotion" GTM GA4 events
      */
-    const HomePageSliderGtmAnalytics4Component = BaseComponent.extend({
+    const HomePageSliderGtmAnalytics4Component = BaseEmbeddedListGtmComponent.extend({
         /**
          * @property {Object}
          */
-        options: _.extend({}, BaseComponent.prototype.options, {
-            nameSelector: '[data-role="slide-content"] h2'
+        options: _.extend({}, BaseEmbeddedListGtmComponent.prototype.options, {
+            nameSelector: '[data-role="slide-item-link"] img'
         }),
 
         /**
@@ -33,7 +33,7 @@ define(function(require) {
          */
         _getModel: function($item) {
             return {
-                name: $item.find(this.options.nameSelector).text().trim()
+                name: $item.find(this.options.nameSelector).attr('alt')
             };
         },