+++ src/Oro/Bundle/GoogleTagManagerBundle/Resources/public/js/app/components/products-datagrid-gtm-component.js
@@ -30,9 +30,11 @@ define(function(require) {
         }),
 
         /**
-         * @property {Boolean}
+         * @property {boolean}
          */
-        _gtmReady: false,
+        get _gtmReady() {
+            return mediator.execute({name: 'gtm:data-layer-manager:isReady', silent: true}) || false;
+        },
 
         /**
          * @property {jQuery.Element}
@@ -69,10 +71,10 @@ define(function(require) {
         delegateListeners: function() {
             ProductsDatagridGtmComponent.__super__.delegateListeners.call(this);
 
-            mediator.once('gtm:data-layer-manager:ready', this._onGtmReady, this);
-
             // Both click and mouseup needed to be able to track both left and middle buttons clicks.
-            this.$datagridEl.on('click mouseup', this.options.productSelector + ' a', this._onClick.bind(this));
+            this.$datagridEl.on(`click.${this.cid} mouseup.${this.cid}`,
+                this.options.productSelector + ' a', this._onClick.bind(this));
+
 
             this.listenTo(this.productsDatagridComponent.grid, 'content:update', this._onView.bind(this));
         },
@@ -101,7 +103,7 @@ define(function(require) {
         },
 
         _onGtmReady: function() {
-            this._gtmReady = true;
+            // @deprecated
         },
 
         /**
@@ -204,9 +206,7 @@ define(function(require) {
                 return;
             }
 
-            mediator.off('gtm:data-layer-manager:ready', this._onGtmReady, this);
-
-            this.$datagridEl.off('click mouseup', this.options.productSelector, this._onClick.bind(this));
+            this.$datagridEl.off(`.${this.cid}`);
 
             ProductsDatagridGtmComponent.__super__.dispose.call(this);
         }